Project/accounts/filters.py
from django_filters import FilterSet
from .models import Post


class PostFilter(FilterSet):
    class Meta:
        model = Post
        fields = {
            'title': ['icontains'],
            'creation_date': ['gte'],
            'author': ['exact'],
        }
